requirements.txt
alembic==1.18.4
bcrypt==5.0.0
blinker==1.9.0
cachelib==0.13.0
click==8.3.1
colorama==0.4.6
coverage==7.13.4
Deprecated==1.3.1
Flask==3.1.3
Flask-Caching==2.3.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.1.1
flask-marshmallow==1.4.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
gunicorn==25.1.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.2.2
ordered-set==4.1.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-flask==1.3.0
python-dotenv==1.2.2
redis==7.3.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
wrapt==2.1.2
